+++ packages/client/src/rtc/__tests__/mocks/webrtc.mocks.ts
@@ -3,6 +3,7 @@ import { vi } from 'vitest';
 const RTCPeerConnectionMock = vi.fn((): Partial<RTCPeerConnection> => {
   return {
     addEventListener: vi.fn(),
+    addIceCandidate: vi.fn(),
     removeEventListener: vi.fn(),
     getTransceivers: vi.fn(),
     addTransceiver: vi.fn(),
@@ -14,15 +15,21 @@ const RTCPeerConnectionMock = vi.fn((): Partial<RTCPeerConnection> => {
     setRemoteDescription: vi.fn().mockResolvedValue({}),
     close: vi.fn(),
     connectionState: 'connected',
+    signalingState: 'stable',
     getReceivers: vi.fn(),
+    getSenders: vi.fn(),
+    removeTrack: vi.fn(),
   };
 });
 vi.stubGlobal('RTCPeerConnection', RTCPeerConnectionMock);
 
 const MediaStreamMock = vi.fn((): Partial<MediaStream> => {
   return {
-    getTracks: vi.fn(),
+    getTracks: vi.fn().mockReturnValue([]),
     addTrack: vi.fn(),
+    removeTrack: vi.fn(),
+    getAudioTracks: vi.fn().mockReturnValue([]),
+    getVideoTracks: vi.fn().mockReturnValue([]),
   };
 });
 vi.stubGlobal('MediaStream', MediaStreamMock);
@@ -31,10 +38,13 @@ const MediaStreamTrackMock = vi.fn((): Partial<MediaStreamTrack> => {
   return {
     addEventListener: vi.fn(),
     removeEventListener: vi.fn(),
-    getSettings: vi.fn(),
+    getSettings: vi.fn().mockReturnValue({ width: 1280, height: 720 }),
     stop: vi.fn(),
+    clone: vi.fn(),
     readyState: 'live',
     kind: 'video',
+    enabled: true,
+    id: crypto.randomUUID(),
   };
 });
 vi.stubGlobal('MediaStreamTrack', MediaStreamTrackMock);
@@ -49,6 +59,7 @@ const RTCRtpTransceiverMock = vi.fn((): Partial<RTCRtpTransceiver> => {
       setParameters: vi.fn(),
     },
     setCodecPreferences: vi.fn(),
+    mid: '',
   };
 });
 vi.stubGlobal('RTCRtpTransceiver', RTCRtpTransceiverMock);
@@ -69,3 +80,12 @@ const RTCRtpReceiverMock = vi.fn((): Partial<typeof RTCRtpReceiver> => {
   };
 });
 vi.stubGlobal('RTCRtpReceiver', RTCRtpReceiverMock);
+
+const RTCRtpSenderMock = vi.fn((): Partial<typeof RTCRtpSender> => {
+  return {
+    getCapabilities: vi.fn(),
+    // @ts-ignore
+    track: vi.fn(),
+  };
+});
+vi.stubGlobal('RTCRtpSender', RTCRtpSenderMock);